Clone of the main dir (1/21) - !Unactive!
This is a temporary clone of main directory. Main purpose is just to replace some basic APIs temporarily to run projects, but it's not saving. As things go back to normal please use standard lstv.ml if possible.

I'll update this later. Currently this snap contains a lot of old unchanged things that basically serve no purpose, ill remove them soon, and that change will probably be aplied to the main directory.